Massive overhaul to the hooks system.  We no longer have a behemoth hooks module that holds all the hooks.  Nor do we have a hooks manager. Instead, we have a passive registration hooks system.  Hooks are created when they're used--the spamming controls their functionality.
Hooks consist of a PriorityQueue and that's it now.

Also changed is how we pass arguments through hooks.  We no longer pass
arguments via a tuple, but instead through a map.  This makes it much
clearer as to what's going on as well as enables us to add additional
arguments in the future without having to worry about the order of the
arguments.

I also made some minor changes to the scripts and the INSTALL file.