Added the config module.  I think this is where all the configuration management pieces will go.
Overhauled all the modules to use the new config module rather than
__init__ (which was a terrible idea--even as a temporary solution).
Removed the StartupEvent and transferred all the pieces to the engine
module.  The engine now has the "main" method which handles all the
command line parsing, figures everything out, instantiates a ui,
imports all the plugins, and then starts the engine.

Switched things around so the engine event loop is now spun off as
a thread an the ui's get the main thread of control.

I actually tested the textui and the tkui this time around.  I've been
sitting on these changes for a day now and everything seems ok.